Add files via upload
This project demonstrates how to create a simple Piano using Arduino. The project uses 8 pushbuttons to generate different musical notes through a piezo buzzer. When a button is pressed, the Arduino plays a corresponding tone on the buzzer, simulating a piano keyboard. This project is a fun and educational way to learn about sound generation using Arduino and the working of basic electronic components like pushbuttons and buzzers