Add files via upload
The Automatic Hand Sanitizer Dispenser is a system that uses an ultrasonic sensor to detect the presence of a hand at a certain distance (typically 20 cm or less). Once the sensor detects a hand, it triggers a servo motor to release a certain amount of sanitizer. This project aims to provide a contactless solution for sanitizing hands, ensuring hygiene and convenience.